requirements.txt
flake8==3.7.9
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
mypy==0.770
psycopg2-binary==2.8.6
pytest-flask==1.0.0
requests==2.23.0
shortuuid==1.0.1
SQLAlchemy==1.3.15
python-dotenv==0.15.0
pandas==1.2.0
flask-redoc==0.2.1 